[jwies/review-budget-cap] review: split the credit-cap tests into credit-cap.test.ts (router.test.ts hit the max-lines lint limit)
